Кочкин Георгий Задание 2-8

+++ 02-java-types/task08/src/com/example/task08/Task08.java
@@ -4,7 +4,7 @@ public class Task08 {
 
     public static boolean solution() {
         int x = 100;
-        double x = Double.POSITIVE_INFINITY;
+        float x = Float.POSITIVE_INFINITY;
         return x == x + 1;
     }
